Add a start of a resource manager. Maybe it should be splitted in two parts (reference counting and resource manager...). For the moment, only shaders and textures are resources.

+++ OpenMe/trunk/OpenMe/src/utils.c
@@ -45,7 +45,7 @@ unsigned int omeTypeToGL(omeType type)
     case OME_FLOAT:     return GL_FLOAT;
     case OME_DOUBLE:    return GL_DOUBLE;
     default:
-        return OME_FAILURE;
+        return (unsigned int)OME_FAILURE;
     }
 }
